@@ -43,7 +43,7 @@
         "phpunit/phpunit": "<9.6",
         "symfony/property-access": "<5.3.0",
         "symfony/property-info": "<5.3.0",
-        "symfony/serializer": "<5.3.1"
+        "symfony/serializer": "<5.4"
     },
     "autoload": {
         "psr-4": {
